track2_backend - Copy/pyproject.toml
[project]
name = "national-agentic-ai-hackathon-2025-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Fasih", email = "161964251+HafizFasih@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "openai-agents>=0.3.2",
    "pandas>=2.3.2",
    "pywa>=3.1.1",
    "rich>=14.1.0",
    "supabase>=2.20.0",
]

[project.scripts]
national-agentic-ai-hackathon-2025-backend = "national_agentic_ai_hackathon_2025_backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
